fix(pr257-review): address 9 Cx1 findings from PR #257 review
F1 (P0, migrations 012): ADD COLUMN IF NOT EXISTS — avoid replay failure.
F2 (P1, observability): isHex reject uppercase hex — align with W3C
    Trace Context Level 2 §2.2.3 (lowercase hex only).
F3 (P1, observability_test): add all-zero trace_id, all-zero span_id,
    version "ff", and uppercase hex traceparent rejection cases.
F4 (P1, rules_wrapper_test): add selector-expr pkg.Type{} accepted
    case for FMT-19.
F5 (P1, rabbitmq/subscriber): document ObservabilityContextMiddleware
    injection requirement — callers bypassing bootstrap must install it
    explicitly or trace restore is a silent no-op.
F6 (P1, observability_test): cover "ctx has empty-string reserved key
    gets overwritten" boundary case.
F7 (P2, outbox/metadata.go): extract CloneMetadata from observability.go
    to a dedicated metadata.go — semantic cohesion.
F8 (P2, observability.go): move ref: comment into ObservabilityMetadata
    godoc so it is not mistaken for package-level documentation.
F9 (P2, rules_wrapper.go): rewrite isPlainStructCompositeType godoc to
    describe the actual nil-Type defensive scenario.

Refs: PR #257 reviewer findings F1-F9 (all Cx1, all IN_SCOPE).
